requirements.txt
paddleocr
paddlepaddle
pillow
python-dotenv
langchain_groq
fastapi
uvicorn
python-multipart